requirements.txt
FlagEmbedding==1.2.11
peft==0.13.2
chromadb==0.5.13
langchain-text-splitters==0.3.0
loguru==0.7.2
pyyaml==6.0.2
fastapi==0.115.2
uvicorn==0.31.1